🚚 flux-backend => move 'zip_project' in flux-backend
