+++ youwol_flux_backend/__init__.py
@@ -1,4 +1,4 @@
 from .configurations import Configuration, Constants, Dependencies
-from .utils import init_resources
+from .utils import init_resources, zip_project
 from .router import get_router
 from .root_paths import *
